Add tests for the scion-pki cert validate flag
Add tests for the `scion-pki cert validate` command and the flag `subject`
